requirements.txt
portalocker>=2.8
requests>=2.31
python-dotenv>=1.0
playwright>=1.41
